Fixed interface names per Symfony code standards
