+++ app/bundles/EmailBundle/Swiftmailer/Transport/InterfaceTokenTransport.php
@@ -14,32 +14,10 @@ namespace Mautic\EmailBundle\Swiftmailer\Transport;
 /**
  * Tokenized Transport Support.
  *
+ * @deprecated 2.12.0 to be removed in 3.0; use TokenTransportInterface instead
+ *
  * Interface InterfaceTokenTransport
  */
-interface InterfaceTokenTransport
+interface InterfaceTokenTransport extends TokenTransportInterface
 {
-    /**
-     * Function required to check that $this->message is instanceof MauticMessage, return $this->message->getMetadata() if it is and array() if not.
-     *
-     * @return mixed
-     */
-    public function getMetadata();
-
-    /**
-     * Return the max number of to addresses allowed per batch.  If there is no limit, return 0.
-     *
-     * @return int
-     */
-    public function getMaxBatchLimit();
-
-    /**
-     * Get the count for the max number of recipients per batch.
-     *
-     * @param \Swift_Message $message
-     * @param int            $toBeAdded Number of emails about to be added
-     * @param string         $type      Type of emails being added (to, cc, bcc)
-     *
-     * @return mixed
-     */
-    public function getBatchRecipientCount(\Swift_Message $message, $toBeAdded = 1, $type = 'to');
 }
